# Upgrade dependencies, mainly lexical for better performance