--- a/crates/jsdoc/Cargo.toml
+++ b/crates/jsdoc/Cargo.toml
@@ -10,14 +10,14 @@ version = "0.61.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-nom = "5.1.2"
+nom = "7.1.0"
 serde = {version = "1", features = ["derive"]}
 swc_atoms = {version = "0.2", path = "../swc_atoms"}
 swc_common = {version = "0.17.0", path = "../swc_common"}
 
 [dev-dependencies]
 anyhow = "1"
-dashmap = "4.0.2"
+dashmap = "5.1.0"
 swc_ecma_ast = {version = "0.69.0", path = "../swc_ecma_ast"}
 swc_ecma_parser = {version = "0.92.0", path = "../swc_ecma_parser"}
 testing = {version = "0.18.0", path = "../testing"}