--- a/crates/node_macro_deps/Cargo.toml
+++ b/crates/node_macro_deps/Cargo.toml
@@ -11,5 +11,5 @@ version = "0.1.0"
 proc-macro = true
 
 [dependencies]
-proc-macro2 = "=1.0.32"
-quote = "=1.0.10"
+proc-macro2 = "1"
+quote = "1"