--- a/crates/preset_env_base/Cargo.toml
+++ b/crates/preset_env_base/Cargo.toml
@@ -13,9 +13,9 @@ version = "0.2.0"
 ahash = "0.7.4"
 anyhow = "1"
 browserslist-rs = "=0.9.0"
-dashmap = "4.0.2"
+dashmap = "5.1.0"
 from_variant = {version = "0.1.3", path = "../from_variant"}
-once_cell = "1.9.0"
+once_cell = "1.10.0"
 semver = {version = "1.0.4", features = ["serde"]}
 serde = {version = "1", features = ["derive"]}
 st-map = "0.1.2"