--- a/crates/swc_bundler/Cargo.toml
+++ b/crates/swc_bundler/Cargo.toml
@@ -29,13 +29,13 @@ default = []
 [dependencies]
 ahash = "0.7"
 anyhow = "1"
-crc = "1.8"
-dashmap = {version = "4.0.2", optional = true}
+crc = "2.1.0"
+dashmap = {version = "5.1.0", optional = true}
 indexmap = "1.6"
 is-macro = "0.2.0"
-once_cell = "1.9.0"
+once_cell = "1.10.0"
 parking_lot = "0.12.0"
-petgraph = "0.5"
+petgraph = "0.6"
 radix_fmt = "1"
 rayon = {version = "1", optional = true}
 relative-path = "1.2"
@@ -59,7 +59,7 @@ hex = "0.4"
 ntest = "0.7.2"
 path-clean = "=0.1.0"
 reqwest = {version = "0.11.4", features = ["blocking"]}
-sha-1 = "0.9"
+sha-1 = "0.10"
 swc_ecma_loader = {version = "0.29.0", path = "../swc_ecma_loader", features = ["node", "cache"]}
 swc_ecma_minifier = {version = "0.88.0", path = "../swc_ecma_minifier"}
 swc_ecma_transforms_react = {version = "0.92.0", path = "../swc_ecma_transforms_react"}