--- a/crates/swc_cached/Cargo.toml
+++ b/crates/swc_cached/Cargo.toml
@@ -14,7 +14,7 @@ version = "0.1.1"
 ahash = "0.7.6"
 anyhow = "1.0.55"
 dashmap = "5.1.0"
-once_cell = "1.9.0"
+once_cell = "1.10.0"
 regex = "1.5.4"
 serde = "1.0.136"
 swc_atoms = {version = "0.2.9", path = "../swc_atoms"}