--- a/crates/swc_cli/Cargo.toml
+++ b/crates/swc_cli/Cargo.toml
@@ -14,23 +14,23 @@ path = "./src/main.rs"
 
 [dependencies]
 anyhow = "1.0.53"
-clap = { version = "3.1.0", features = ["derive", "wrap_help"] }
-walkdir = "2"
-rayon = "1"
-swc = { version = "0.149.0", path = "../swc" }
-swc_common = { version = "0.17.5", path = "../swc_common" }
-relative-path = "1.6.1"
-serde = { version = "1", features = ["derive"] }
-serde_json = { version = "1", features = ["unbounded_depth"] }
 atty = "0.2.14"
+clap = {version = "3.1.0", features = ["derive", "wrap_help"]}
+rayon = "1"
+relative-path = "1.6.1"
+serde = {version = "1", features = ["derive"]}
+serde_json = {version = "1", features = ["unbounded_depth"]}
+swc = {version = "0.149.0", path = "../swc"}
+swc_common = {version = "0.17.5", path = "../swc_common"}
 swc_trace_macro = {version = "0.1.0", path = "../swc_trace_macro"}
+tracing = "0.1.32"
 tracing-chrome = "0.4.0"
 tracing-futures = "0.2.5"
-tracing-subscriber = { version = "0.3.9", features = ["env-filter"] }
-tracing = "0.1.32"
+tracing-subscriber = {version = "0.3.9", features = ["env-filter"]}
+walkdir = "2"
 
 [dependencies.path-absolutize]
-version = "3.0.11"
 features = ["once_cell_cache"]
+version = "3.0.11"
 
 [features]