--- a/crates/swc_common/Cargo.toml
+++ b/crates/swc_common/Cargo.toml
@@ -30,12 +30,12 @@ arbitrary = {version = "1", optional = true, features = ["derive"]}
 ast_node = {version = "0.7.5", path = "../ast_node"}
 atty = {version = "0.2", optional = true}
 better_scoped_tls = {version = "0.1.0", path = "../better_scoped_tls"}
-cfg-if = "0.1.2"
+cfg-if = "1.0.0"
 debug_unreachable = "0.1.1"
 either = "1.5"
 from_variant = {version = "0.1.3", path = "../from_variant"}
-num-bigint = "0.2"
-once_cell = "1.9.0"
+num-bigint = "0.4"
+once_cell = "1.10.0"
 parking_lot = {version = "0.12.0", optional = true}
 rkyv = {version = "0.7.28", optional = true}
 rustc-hash = "1.1.0"