… Cargo.lock
… crates/swc_css_parser/Cargo.toml
… crates/swc_ecma_ast/Cargo.toml
… crates/swc_ecma_codegen/Cargo.toml
… crates/swc_ecma_dep_graph/Cargo.toml
… crates/swc_ecma_ext_transforms/Cargo.toml
… crates/swc_ecma_lints/Cargo.toml
… crates/swc_ecma_loader/Cargo.toml
… crates/swc_ecma_minifier/Cargo.toml
… crates/swc_ecma_parser/Cargo.toml
… crates/swc_ecma_parser/benches/files/three-0.138.3.js
… crates/swc_ecma_preset_env/Cargo.toml
… crates/swc_ecma_transforms/Cargo.toml
… crates/swc_ecma_transforms_base/Cargo.toml
… crates/swc_ecma_transforms_compat/Cargo.toml
… crates/swc_ecma_transforms_optimization/Cargo.toml
… crates/swc_ecma_transforms_proposal/Cargo.toml
… crates/swc_ecma_transforms_react/Cargo.toml
… crates/swc_ecma_transforms_testing/Cargo.toml
… crates/swc_ecma_utils/Cargo.toml
… crates/swc_ecma_visit/Cargo.toml
… crates/swc_eq_ignore_macros/Cargo.toml
… crates/swc_estree_compat/Cargo.toml
… crates/swc_fast_graph/Cargo.toml
… crates/swc_graph_analyzer/Cargo.toml
… crates/swc_macros_common/Cargo.toml
… crates/swc_node_bundler/Cargo.toml
… crates/swc_node_comments/Cargo.toml
… crates/swc_plugin_runner/Cargo.toml
… crates/testing/Cargo.toml