## Before

### Lexer
test three         ... bench:  14,350,033 ns/iter (+/- 767,897) = 82 MB/s

### Parser:
test three         ... bench:  27,030,841 ns/iter (+/- 765,058) = 43 MB/s

## After

### Lexer
test three         ... bench:  14,089,720 ns/iter (+/- 667,112) = 84 MB/s

### Parser
test three         ... bench:  26,409,095 ns/iter (+/- 1,086,535) = 44 MB/s